remove the unused codes for pprof
