+++ go.mod
@@ -40,7 +40,6 @@ require (
 	github.com/twmb/murmur3 v1.1.5
 	github.com/wacul/ptr v1.0.0 // indirect
 	golang.org/x/tools v0.1.0
-	google.golang.org/genproto v0.0.0-20200526211855-cb27e3aa2013 // indirect
 	google.golang.org/protobuf v1.26.0
 	honnef.co/go/tools v0.0.1-2020.1.6
 )
