+++ pkg/agent/pprof/map.go
@@ -68,8 +68,7 @@ Search:
 	if len(m.freeStk) < len(stk) {
 		m.freeStk = make([]uintptr, 1024)
 	}
-	// Limit cap to prevent append from clobbering freeStk.
-	e.stk = m.freeStk[:len(stk):len(stk)]
+	e.stk = m.freeStk[:len(stk)]
 	m.freeStk = m.freeStk[len(stk):]
 
 	for j := range stk {
